move handleDatasetAdd to mobile dropdown folder
